pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sc-heimdall"
dynamic = ["version"]

authors = [
  { name="Remy Liu"},
  { name="Nicholas Ho"},
  { name="Shahul Alam", email="alam.shahul@gmail.com" },
]

dependencies = [  # TODO: check which are necessary
    "accelerate",
    "anndata>=0.11.0",
    "hydra-core",
    "numpy",
    "omegaconf",
    "packaging",
    "pandas",
    "wandb",
    "tqdm",
    "scanpy[leiden]",
    "scikit-learn",
    "scipy",
    "torch==2.5.*",
    "torchmetrics",
    "transformers",
    "datasets",
    "einops",
    "mygene",
    "awkward",
]
description = "Heimdall: A Comprehensive Paradigm for Evaluating Single-Cell Representations within Foundational Models"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
flash = [
    "flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.4.post1/flash_attn-2.7.4.post1+cu12torch2.5cxx11abiFALSE-cp312-cp312-linux_x86_64.whl ; sys_platform == 'linux' and python_version == '3.12'",
]

[tool.hatch.build.targets.sdist]
include = [
  "Heimdall/**/*.py",
  "Heimdall/config/**",
]

[tool.hatch.build.targets.wheel]
include = [
  "Heimdall/**/*.py",
  "Heimdall/config/**",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "Heimdall/__about__.py"

[tool.flake8]
max-line-length = 120
count = true
extend-ignore = [
    "E203",  # conflicting with black style slicing
    "D107",  # only need docstring at class level, no need at __init__
]
exclude = [
    ".nox",
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.pyc",
    "*.egg-info",
    ".eggs",
]

[tool.black]
line-length = 120
preview = true

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
markers = [
    "integration: marks tests for entire Heimdall pipeline"
]

testpaths = [
    "tests/"
]

# Dev requirements
[dependency-groups]
dev = [
    "ipdb",
    "ipython",
    "nox",
    "pre-commit",
    "pytest",
    "python-dotenv"
]
